Available Models Table

Table containing the available models to be fit using fit_devmodels(). These models come from two other packages: devRate and rTPC .

Usage

data("available_models")

Format

available_models

A data.frame/tibble with 13 rows and 6 columns:

model_name

Model name to be used within fit_devmodels().

package

names of the packages used by fit_devmodels() to obtain appropriate start values for the user-provided data. When the package is rTPC package, start values are automatically computed with rTPC::get_start_vals(), which in turn relies on nls.multstart::nls_multstart(). When the package is devRate package, iterative starting values are computed using nls.multstart::nls_multstart(), using the parameters published in devRate::devRateEqStartVal() as first attempts to iterate. As an exception, if model_name == "briere1", generic starting values are provided and advised to the user due to the unrealistic value of some parameters in the devRate data set.

source_model_name

name of the function in the source packages rTPC and devRate.

formula, working_formula, n_params

formulas used for model fitting.

Source

Rebaudo, F., Struelens, Q., and Dangles, O. (2018). Modelling temperature-dependent development rate and phenology in arthropods: The devRate package for r. Methods Ecol Evol. 9: 1144-1150. https://doi.org/10.1111/2041-210X.12935Padfield, D., O´Sullivan, H., and Pawar, S., (2021). rTPC and nls.multstart: a new pipeline to fit thermal performance curves in r. Methods Ecol Evol. 12: 1138-1143 https://doi.org/10.1111/2041-210X.13585.